requirements.txt
requests>=2.31
python-dotenv>=1.0
supabase>=2.0
anthropic>=0.40
pandas>=2.0
xlrd>=2.0
openpyxl>=3.1
rapidfuzz>=3.6
